requirements.txt
numpy
customtkinter
tkinter 
pillow